% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_query_string.R
\name{check_numeric}
\alias{check_numeric}
\title{Check numeric}
\usage{
check_numeric(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character string or a numeric. it depends on the pattern of the given string.
}
\description{
Convert a given string into a numeric value if the string was supposed to be a number.
}

