% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_hostname}
\alias{get_hostname}
\title{Get host name from the given url}
\usage{
get_hostname(url)
}
\arguments{
\item{url}{A length-one character vector}
}
\value{
the host name A length-one character vector
}
\description{
Get host name from the given url
}

