% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_composer.R
\name{get_query_part}
\alias{get_query_part}
\title{Extract query part from the given url string}
\usage{
get_query_part(url)
}
\arguments{
\item{url}{A length-one character vector}
}
\value{
the query part of the url A length-one character vector
}
\description{
Extract query part from the given url string
}

