% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_query_string.R
\name{get_query_string}
\alias{get_query_string}
\title{Extract query string}
\usage{
get_query_string(url)
}
\arguments{
\item{url}{A character string.}
}
\value{
A character string that represents the query string.
}
\description{
Extracts query string from a given url
}

