% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{has_query_params}
\alias{has_query_params}
\title{Test if a given url has query paramters or not}
\usage{
has_query_params(url)
}
\arguments{
\item{url}{A length-one character vector}
}
\value{
A length-one logical vector
}
\description{
Test if a given url has query paramters or not
}

