% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse}
\alias{parse}
\title{Parse a link header string with multiple urls & rels then convert it into a list}
\usage{
parse(link_header_string)
}
\arguments{
\item{link_header_string}{A length-one character}
}
\value{
the parsed link header A list
}
\description{
Parse a link header string with multiple urls & rels then convert it into a list
}
\examples{
parse('<http://example.com?limit=50&order=desc>; rel="meta", <http://example.com?limit=100&order=asc>; rel="prev", <http://example.com>; rel="last"')
}

