% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_link.R
\name{parse_link}
\alias{parse_link}
\title{Parse a link}
\usage{
parse_link(x)
}
\arguments{
\item{x}{A character string indicating \code{link}/\code{rel} information. (E.g <'www.example.com>; rel="meta"')}
}
\value{
A list keyed by the \code{rel} property that contains information about the link
}
\description{
Parses a single link/rel information
}

