% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_links.R
\name{parse_links}
\alias{parse_links}
\title{Parse links}
\usage{
parse_links(x)
}
\arguments{
\item{x}{A character string indicating \code{link}s/\code{rel}s information. (E.g '<www.example.com/first>; rel="first", <www.example.com/second>; rel="second"')}
}
\value{
A list keyed by the \code{rel} properties that contains information about each link
}
\description{
Parses a link header text into an R list that contains each contained link
}
\examples{
parse_links('<http://example.com/first>; rel="meta", <http://example.com>; rel="home"')
}

