% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_one}
\alias{parse_one}
\title{Parse a link header string with single url & rel then convert it into a list}
\usage{
parse_one(link_header_string)
}
\arguments{
\item{link_header_string}{A length-one character}
}
\value{
the parsed link header A list
}
\description{
Parse a link header string with single url & rel then convert it into a list
}

