% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_query_string.R
\name{parse_query_string}
\alias{parse_query_string}
\title{Parse query string}
\usage{
parse_query_string(url)
}
\arguments{
\item{url}{A character string.}
}
\value{
A list containing a representation of query paramters.
}
\description{
Parses query string into an R list.
}

