% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseLinks.R
\docType{package}
\name{ParseLinks}
\alias{ParseLinks}
\alias{ParseLinks-package}
\title{ParseLinks: A simple link header parser for R}
\description{
The ParseLinks package provides a parser function for link header strings. 
By using `parse_links` function, you can get an R list for each contained link in the link header strings.
}

