% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_composer.R
\name{query_composer}
\alias{query_composer}
\title{Compose query params list from the given query paramter string}
\usage{
query_composer(url)
}
\arguments{
\item{url}{A length-one character vector}
}
\value{
the query paramters A list
}
\description{
Compose query params list from the given query paramter string
}

