% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_link.R
\name{remove_query_params}
\alias{remove_query_params}
\title{Remove query paramter part from a url}
\usage{
remove_query_params(url)
}
\arguments{
\item{url}{character string.}
}
\value{
The query parameter removed url. A character string.
}
\description{
Remove query paramter part from a url
}

