% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter.R
\name{splitter}
\alias{splitter}
\title{Splitter}
\usage{
splitter(x)
}
\arguments{
\item{x}{A character string}
}
\value{
A character vector
}
\description{
Splits a text string by comma then returns a character vector
}

