% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_composer.R
\name{cast_numeric}
\alias{cast_numeric}
\title{Convert to numeric if the given string represents a number}
\usage{
cast_numeric(x)
}
\description{
Convert to numeric if the given string represents a number
}

