% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_links.R
\name{has_query_params}
\alias{has_query_params}
\title{Test if a given url has query paramters or not}
\usage{
has_query_params(url)
}
\arguments{
\item{url}{A character string}
}
\value{
A logical
}
\description{
Test if a given url has query paramters or not
}

