% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter.R
\name{splitter}
\alias{splitter}
\title{Parse a link header text then convert the text into a character vector 
such like c('<url..>; rel=".."', '<url..>; rel=".."')}
\usage{
splitter(link_header_string)
}
\arguments{
\item{link_header_string}{A length-one character vector}
}
\value{
A character vector
}
\description{
Parse a link header text then convert the text into a character vector 
such like c('<url..>; rel=".."', '<url..>; rel=".."')
}

